Rename MenuInterface file name
