+++ Headers/UserInterface/AwardsInterface.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "UserInterface/MenuInterface.h"
+#include "UserInterface/PromptInterface.h"
 
 #define INSERT_COIN_OPTION "Insert Coin"
