+++ Headers/UserInterface/LoadingInterface.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "UserInterface/MenuInterface.h"
+#include "UserInterface/PromptInterface.h"
 
 class LoadingInterface final : public PromptInterface
 {
