+++ Headers/UserInterface/PostgameInterface.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "UserInterface/MenuInterface.h"
+#include "UserInterface/PromptInterface.h"
 
 #define BACK_TO_MAIN_MENU_OPTION "Return to Main Menu"
 #define AWARDS_OPTION "Awards"
