+++ Headers/UserInterface/PregameInterface.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "UserInterface/MenuInterface.h"
+#include "UserInterface/PromptInterface.h"
 
 #define START_GAME_OPTION "Start Game"
 #define BACK_OPTION "Back"
