--- Source/UserInterface/MenuInterface.cpp
@@ -1,10 +1,0 @@
-#include "UserInterface/MenuInterface.h"
-
-PromptInterface::PromptInterface(
-    vector<pair<float, float>> componentScaling,
-    vector<pair<float, float>> componentTranslating)
-    : UserInterface(
-        componentScaling, componentTranslating)
-{
-    /* TODO */
-}
