+++ Source/UserInterface/PromptInterface.cpp
@@ -1,0 +1,10 @@
+#include "UserInterface/PromptInterface.h"
+
+PromptInterface::PromptInterface(
+    vector<pair<float, float>> componentScaling,
+    vector<pair<float, float>> componentTranslating)
+    : UserInterface(
+        componentScaling, componentTranslating)
+{
+    /* TODO */
+}
